// filter-frontend/src/app/enums/comparison-type.enum.ts
export enum ComparisonType {
    EQUAL_TO = 'equal_to',
    NOT_EQUAL_TO = 'not_equal_to',
    GREATER_THAN = 'greater_than',
    LESS_THAN = 'less_than',
    GREATER_THAN_OR_EQUAL_TO = 'greater_than_or_equal_to',
    LESS_THAN_OR_EQUAL_TO = 'less_than_or_equal_to',
    STRING_EQUAL_TO = 'string_equal_to',
    STRING_NOT_EQUAL_TO = 'string_not_equal_to',
    STRING_CONTAINS = 'string_contains',
    STRING_STARTS_WITH = 'string_starts_with',
    STRING_ENDS_WITH = 'string_ends_with',
    STRING_MATCHES = 'string_matches',
    DATE_EQUAL_TO = 'date_equal_to',
    DATE_NOT_EQUAL_TO = 'date_not_equal_to',
    DATE_BEFORE = 'date_before',
    DATE_AFTER = 'date_after',
    DATE_ON_OR_BEFORE = 'date_on_or_before',
    DATE_ON_OR_AFTER = 'date_on_or_after'
}